# python-flask/swagger/swagger.yaml
---
swagger: "2.0"
info:
  version: ""
  title: ""
schemes:
- "http"
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
  /api/1.0/create:
    post:
      tags:
      - "PostItService"
      summary: "PostItService.createPostItTable"
      operationId: "controllers.default_controller.create_post_it_table"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/postitCreatePostItTable"
      responses:
        default:
          description: "Description"
          schema:
            $ref: "#/definitions/postitCreatePostItTableResponse"
      x-tags:
      - tag: "PostItService"
  /api/1.0/delete:
    post:
      tags:
      - "PostItService"
      summary: "PostItService.deleteApiKey"
      operationId: "controllers.default_controller.delete_api_key"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/postitDeleteApiKey"
      responses:
        default:
          description: "Description"
          schema:
            $ref: "#/definitions/postitDeleteApiKeyResponse"
      x-tags:
      - tag: "PostItService"
  /api/1.0/post:
    post:
      tags:
      - "PostItService"
      summary: "PostItService.postIt"
      operationId: "controllers.default_controller.post_it"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/postitPostIt"
      responses:
        default:
          description: "Description"
          schema:
            $ref: "#/definitions/postitPostItResponse"
      x-tags:
      - tag: "PostItService"
  /api/1.0/regenerate:
    post:
      tags:
      - "PostItService"
      summary: "PostItService.regenerateApiKey"
      operationId: "controllers.default_controller.regenerate_api_key"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/postitRegenerateApiKey"
      responses:
        default:
          description: "Description"
          schema:
            $ref: "#/definitions/postitRegenerateApiKeyResponse"
      x-tags:
      - tag: "PostItService"
  /api/1.0/update:
    post:
      tags:
      - "PostItService"
      summary: "PostItService.updatePostItTable"
      operationId: "controllers.default_controller.update_post_it_table"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/postitUpdatePostItTable"
      responses:
        default:
          description: "Description"
          schema:
            $ref: "#/definitions/postitUpdatePostItTableResponse"
      x-tags:
      - tag: "PostItService"
definitions:
  postitCreatePostItTable:
    type: "object"
    properties:
      path:
        type: "string"
        format: "string"
      schema:
        type: "string"
        format: "string"
      shard_name:
        type: "string"
        format: "string"
  postitCreatePostItTableResponse:
    type: "object"
    properties:
      api_key:
        type: "string"
        format: "string"
  postitDeleteApiKey:
    type: "object"
    properties:
      api_key:
        type: "string"
        format: "string"
  postitDeleteApiKeyResponse:
    type: "object"
  postitPostIt:
    type: "object"
    properties:
      api_key:
        type: "string"
        format: "string"
      event_time:
        type: "integer"
        format: "int64"
      payload:
        type: "string"
        format: "string"
  postitPostItResponse:
    type: "object"
  postitRegenerateApiKey:
    type: "object"
    properties:
      api_key:
        type: "string"
        format: "string"
  postitRegenerateApiKeyResponse:
    type: "object"
    properties:
      api_key:
        type: "string"
        format: "string"
  postitUpdatePostItTable:
    type: "object"
    properties:
      api_key:
        type: "string"
        format: "string"
      path:
        type: "string"
        format: "string"
      schema:
        type: "string"
        format: "string"
  postitUpdatePostItTableResponse:
    type: "object"